+++ packages/types/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Deprecated flare SDK (#2428)
+
 ## [3.6.1] - 2024-05-27
 ### Changed
 - Type for unsafeApi to from Provider to JsonRpcProvider (#291)